optimize InputSelect
